Bounding spheres on objects
We don't want to iterate through all the empty space, so each object
defines a bounding sphere, and we skip ahead until we hit one of these
spheres.

+++ utils.h
@@ -1,7 +1,7 @@
 #ifndef __UTILS_H__
 #define __UTILS_H__
 typedef int i;
-typedef float f;
+typedef double f;
 #define max(_a,_b) ((_a)>(_b)?(_a):(_b))
 #define min(_a,_b) ((_a)<(_b)?(_a):(_b))
 #define posmod(_a,_b) (((_b)+((_a)%(_b)))%(_b))
@@ -10,12 +10,12 @@ typedef float f;
 
 v mirror(v dir,v norm);
 
-float clamp(float a, float min, float max);
+double clamp(double a, double min, double max);
 
-float R();
+double R();
 
-v calculateDirection(float x, float y);
+v calculateDirection(double x, double y);
 
-float intersectSphereDist(v center, float R, v start, v direction);
-bool intersectSphere(v center, float R, v start, v direction);
+double intersectSphereDist(v center, double R, v start, v direction);
+bool intersectSphere(v center, double R, v start, v direction);
 #endif
